Deleting everything and starting over.
I lost momentum on this project when I got distracted with other things. I didn't get very far and wasn't super-excited with Jekyll at the time and have since forgotten everything about it.

I'd rather start over from scratch. I think I'll go with Eleventy because all the cool kids are telling me it's the way to go and I want to be a cool kid. I don't feel too strongly about the tech stack otherwise, and I think the most important technical aspect for a blog like this is to be fast for users. @slightlyoff in particular seems to think it's the way to go, and I'm too much of a sheeple to disagree.