Adds a home page which lists all posts.
I opted to use Liquid for the HTML templating engine because I found its syntax more clear and self-describing, rather than being very symbol focused like Mustache.

This includes a second post to test out sorting with.

I also added dates to the front matter of both posts. These are hard coded rather than relying on file creation date as that can change during deployment and is a little too magical for my taste. Unfortunately this date is in UTC time, so I needed to add a timezone (which is dependent on daylight savings).