Adds a layout for a post.
This constructs a proper web page for each post and includes the title as web as a simple about section. "About the author" was pulled into a separate fragment so it can be shared across pages.

I separated layouts into two groups, "pages" and "fragments". Pages are top-level and should not be included in another template, while fragments are pieces of composable UI meant to be included wherever they are needed.

For example, a page's `layout` frontmatter should always refer to `pages/...` while an `{% include ... %}` statement should always refer to `fragments/...`.