Moves styles into fragments directory.
This more closely ties the styles for a particular fragment with the HTML it is associated with.

The best way I came up with here was for each fragment to include a macro that emits the styles for that fragment. Pages or higher-order fragments which include a fragment should also include its styles in a `<style></style>` tag or another `styles()` macro.

This design allows each fragment to be a little more self-contained and composable.