Update fragment stylest to use `frag-*` class name.
This class namespaces each component with a unique class to scope styles to that particular component. Previously we were using the tag name (`header` or `footer`) which could have unintended side effects of other tags on the page.

Unfortunately this raised the specificity of some footer styles which had some side effects of overriding some default margins. I needed to be much more explicit about not overridding those values to get the desired styling.