Componentizes the Card fragment.
This moves the card styles and markup to a separate fragment for better encapsulation.

Unfortunately, I couldn't find a good way of doing a "block call", that is, a `call` which includes multiple blocks inside it, one for the title and one for the content. Nunjucks doesn't seem to support this. Instead, I opted to use subclasses to provide the styles needed for particular DOM elements in the content. This doesn't quite provide the encapsulation I wanted as it is relatively easy to mis use by including the wrong tag or typo-ing the class name. However, this is probably good enough for the current purposes.