Adds basic themeing to the page.
This uses the [Lora](https://fonts.google.com/specimen/Lora) font as a base with some simple colors. Each of the text colors holds up to text accessibility guidelines and is AAA-compatible according to the DevTools color picker. The one exception is the alt visisted link color, which is intended to be used on the primary blue background. This is only AA compatible, but is still mostly legible and a slight enough purple that a sample size of 1 was able to identify it as a visited link. The same sample also found that a simple underline of white text was sufficient to mark it as a link, so the alt unvisited link color was left as a simple white.

I included some docs with some of the tools I found to identify compatible colors and check their accessibility.