Expands the `<main />` height to grow with the flexbox.
Previously the `<main />` element was only as large as necessary to contain the content. However, if the device viewport was larger, then that extra space would have the default background color (white) rather than the `<main />` element's background color.

This change allows the `<main />` element to expand with the flexbox and fill as much space as needed. In doing so, the background color is correct for any size viewport.