Adds tests for the styles filters file.
This simply tests that files are included, deduplicated, and concatenated directly. I also made a minor modification of normalizing the absolute paths before deduplicating them in order to catch multiple uses of the same file via different paths (for example, `foo.css` and `../www/foo.css`).

The de-duplication change still isn't all that smart. It does not look at the filesystem at all, so it can't track things like symlinks to the same file. This should be good enough for now, and I can expand upon it later if necessary.