Migrates `testing/filters_test.js` to TypeScript.
Unfortunately covariance/contravariance is a bit of a challenge here, since the simple implementations of filters are not technically broad enough to include all the inputs allowed by the type interface. As a simple workaround, I hard-casted it to the specific type, but this should be improved in TS 4.0, when array types can have spreads.