Moves `www/` directory to `src/www/`.
This keeps the frontend codebase in line with the Node code. It also puts all HTML/JS/TS/CSS under `src/`, which separates it from the configuration files outside of there.

Only additional meaningful edits was a minor change to `src/filters/styles.ts` (and tests) to change the root directory.