Adds Netlify configuration.
This includes the `netlify.toml` file, which configures the build and server information. Currently this is a typically 11ty build, but we need to run `npm run build` in order to build TypeScript first.

This also includes a new CI job to deploy, which should run only on the `deploy` branch. This uses the Netlify GitHub action to push a build. I believe Netlify will still pick up the `netlify.toml` file, as there are countless settings there which are not supported by their GitHub action.

Updated the README with the general policy and expectations around deployments. The PR preview feature has not been tested, but will hopefully work.

As with all GitHub actions, it will likely take a few tries before it works as expected, but there's currently no great way of testing until I just go for it.