Adds Git data.
This reads the local repository Git information to get the hash value of the current commit.

This required using `export = ` rather than `export default` in order to be compatible with 11ty's CommonJS usage.

It is apparently also impossible to spy on `child_process.execFile()`, as it cannot be modified or even have a getter property stubbed. To work around this, I simply provide `execFile` as an argument, but default to the real `execFile`. This way it can be stubbed for tests but still works correctly in production.