Adds a `/deployment` page.
This includes the current Git commit as well as a link to the GitHub diff for it. This is mainly useful for confirming the deployed version a user is currently looking at.

This page is unstyled and mostly intended as a debug page. I don't anticipate any "real" users caring about this.