Add `NETLIFY_CMD` and `NETLIFY_DIR`.
Netlify deployment GitHub action seems to say it is successful, but doesn't actually deploy. It doesn't show up in Netlify deployments, and the production site does not appear to change.

I'm guessing that pershaps this does not pick up `netlify.toml`, so I'm setting `NETLIFY_CMD` and `NETLIFY_DIR` as a quick test.