Attempts to use the Netlify CLI to deploy on successful CI run of the `deploy` branch.
This adds a new `Deploy` action to the CI workflow, which triggers the Netlify deploy to prod.

I'm not really sure what was wrong with the GitHub action, it would run successfully but Netlify never showed a build nor would the site actually update.

This also has a side benefit over the Netlify GitHub action, because this one does not actually build the application, it only pushes a pre-built package. This is good so only GitHub actions is responsible for building the application (just one environment to support) and also means faster and less complex deploys. Removed the `command` attribute of Netlify's configuration to be more clear about this and less likely to accidentally build with Netlify in the future.