Fixes broken globstar in `npm run clean`.
NPM scripts use the basic `sh` command, which does not support globstar, so `src/**/*.js` is actually just `src/*/*.js`. This means that it was not properly recursive and would not clean `src/www/_data/git.js`.

Separately, Eleventy apparently reads **all** `*.js` files in the `_data/` directory. As a result, it would try to load `git_test.js` which was leftover from a previous test (due to not being cleaned). It would then fail because Jasmine is not loaded into Eleventy.

The simple fix here is to update clean to actually clean out all recursive JavaScript files. Since we can't use globstar, the best way I found was to use the `find` command with `-delete`. Hopefully that is a standard enough utility that it shouldn't be too big a deal.

Separately, I should look into making 11ty more accomodating of test files in the `_data/` directory, but a quick attempt with `.eleventyignore` was not successful. I'll follow up with that later, this should be good enough to unblock CI for the time being.