Adds `main` branch continuous deployment.
This deploys the `main` branch on push with a simple alias. Unfortunately, the alias `main` appears broken, and deploys are successful but do not appear at that URL. Any other alias works just fine, so I think `main` is just broken since it was already previously used as a branch build. Workaround is to just use `main-preview` instead.