Passes in global data to footer macro.
Apparently, by default Nunjucks macros do not inherit the context of the caller, so they do not have access to global data. This manifested in broken links in the footer.

It is possible to include `with context` on the import, but that was a little too heavyweight for my preference. https://github.com/11ty/eleventy/issues/434

Instead, I opted to just pass in that global data where necessary. This is sort-of exactly the same with extra work for the caller. However it allows me to snoodily say that global variables are an anti-pattern and I avoid using them wherever possible.