Removes `deploy` from PR actions.
Pull requests should not be made to the `deploy` branch, they should only be made to `main`.

I'm slightly concerned that the action will deploy to the main preview on pull request. I don't think this will happen based on the `github.ref` condition, as I think that will not match the `main` branch. Removing `deploy` at least prevents an accidental production deployment, but I'll see if a PR to `main` triggers a preview deployment.