Trims whitespace before period.
This introduces a `-` before the `if`, which tells Nunjucks to trim whitespace before the `if`. This allows the period to immediately follow the commit hash without any spacing between them.

I'm intentionally **not** including the `-` at the end of the `if`, because we want whitespace when "(with local changes)" is included, this way there is a space between the commit and the parentheses. The `-` after the `endif` is sufficient to make sure there is no space between "(with local changes)" and the ending period.