Adds `asserts.ts` file with tests.
This is a simple library to assert that a value is defined. This is most useful for LitElement, where most custom elements are defined with optional properties due to the way the custom elements life cycle works. Any required properties need to be asserted to be defined, where this utility becomes very helpful.