Updates naming conventions in `package.json` scripts and `tsconfig` files.
As we introduce more builds for the browser, service worker, etc., I wanted to be more clear about each build command and exactly what it is doing. This renaming mostly includes:
* `app` -> `11ty`.
* `test` -> `11ty-test`.

The convention here is that basically every "compile target" will have `build:<name>` as well as a `build:<name>-test` and a `test:<name>`. Karma may differ slightly, but this should be a rough pattern to follow for new compile targets.