Moves 11ty code to a new `11ty` directory.
This makes it easier to separate code for different build targets by giving each its own subdirectory under `src/`. This commit moves Eleventy sources to an `11ty` directory and updates the config files to match.

I chose to leave `www/` as a direct dependency of `src/` since it doesn't really have executable JavaScript in the same sense and it mostly maps the serving structure of the final site. I felt that was unique enough to justify a separate directory, even though it will be built as part of Eleventy.