Adds Babel transform to Karma typescript so it can process ESM code for LitElement.
Testing code which uses LitElement would otherwise fail due to it being written in ESM. By adding this transform, Babel processes the `import/export` syntax and converts it be ES5 compatible and work with Karma.

The Babel runtime plugin is necessary because the transformer generates code which expects that runtime to be present.

I also chose to set `noEmitOnError` for the TypeScript build as this made Karma errors a little nicer and just left the TypeScript error at the end of the terminal, which is easier to read.