Adds `dwac-share` component.
This provides sharing functionality via a "share" button which triggers the user agent's share API as well as a "copy" button which simply copies to the clipboard.

This component supports its inputs coming from attributes, so it can be generated with static inputs at build time.

Follow up work to come:
* CSS styling for the component (just the implementation is here for now).
* Direct integration with Twitter when `navigator.share()` is unsupported.
* A small toast message to tell the user that the link was copied to the clipboard.