Adds share component to end of post.
This provides a hook and prompt at the bottom of the post page to remind users they may want to share the post.

I shifted the author down so the share prompt will come first. Most users will not read past the "About the author" section, so I think it is more important for the share component to come first.