Adds filter to minify CSS.
This filter processes and resolves local `@import` statements before minifying the entire bundle.

If CleanCSS reports any errors, then the build also throws the error. Reported warnings are printed to the console.

I needed to make a `clean_css.ts` file in order to easily mock and test the utility. Otherwise it is quite tricky to mock the default export with an unconfigureable property.