Migrates index.html to use the new Clean CSS toolchain.
Style dependencies are shifted into the `index.css` file where Clean CSS can resolve and bundle them.

I opted for optimization level 2, the most optimal form. It doesn't seem to show any problems, so I'll be as strict as possible until I find something becomes a problem.

This does show two warnings for empty properties (`--foo: /**/;`), which is a trick I used in a couple places. I'm leaving the warning alone for now, but I'll need to refactor how that works to get rid of it.