Minifies HTML output with `html-minifier-terser`.
Looking over a few HTML minification competitors, this one seems to provide the smallest outputs and is the one generally recommended by others. Although it seems to be a fork of https://github.com/kangax/html-minifier, where development appears to have stopped for the last year. I think it might have just gotten renamed or changed owners, but the old repo still exists just to add confusion.

I implemented this inline and didn't bother pulling it to a separate module and testing it because it is so simple and highly dependent on the 11ty transform API. It's small enough that I'm not too concerned about the test coverage here.

Unfortunately it seems that most of the options are off by default and the minification option does next to nothing with default options. I went through all the options and enabled just about everything I could. A couple notable ones:
* `collapseInlineTagWhitespace` - I left this off because we actually utilize such whitespace in footer (between the text node and the following link). I think it is reasonable to leave a space for those situations.
* `minifyCSS`/`minifyJS` - I left these off because they are done before their contents are injected into the HTML (JavaScript is external anyways).
* `minifyURLs` - This sounds like a great way to break a link while providing minimal size improvements, so I just left that off. It would also require knowledge of the domain I'm hosted at, which I have deliberately avoided so far to remain portable.

There were also a few like `useShortDoctype` and `removeScriptTypeAttributes` which should really be done in the source code, but I decided to enable them in case one sneaks through.