Adds LitElement literal minifier to Rollup build.
This minifies `html` and `css` template literals using the same tools already leveraged by 11ty. I factored out the configuration to a separate file so it could be used by both 11ty and Rollup. Unfortunately this meant they needed to be written in CommonJS JavaScript, which is a bit a unfortunate. However, some clever usage of `@ts-check` and annotations are good enough to provide type checking in the editor, even if it is not enforced in the actual build.

This should mean that HTML and CSS template literals are minified the same was as 11ty will do it.

This still doesn't really solve the bundle size challenges, as it saves less than 2 KB. Nonetheless, this will be useful going forward, as the number of components and their implementations grow.