Removes extra gutters on the index page header.
This made the header size differ between the index page and post page. The difference is enough to cause a line break on one but not the other for certain devices.

Removing the gutter means the header no longer lines up with the page content, but I think that is a separate problem to solved at the site level, rather than on individual pages.