Adds CSS files as a watch target.
Since these files are loaded via a filter in `*.njk` templates, 11ty isn't smart enough to know that they should be considered source files. Adding this watch target enables it to trigger a live reload whenever styles change.

Unfortunately 11ty does **not** appear to wait for the template compilation to complete. It reloads the browser before the new state has been generated, so nothing is updated. If I manually follow up with a refresh that works, so it seems like 11ty is jumping the gun here. It's recompiling templates based on the CSS change, but the live reload is not appropriately waiting for that template compliation to complete.