Removes duplicate margins from footer.
For whatever reason the index and post pages render this differently. On the post page, the margins overlap, so a top and bottom maring of 5px from adjacent elements results in a total distance of 5px. While on the index page, the margins do not overlap, so the total distance is 10px.

I'm not sure why this difference exists, but the easiest solution is to remove the overlapping margins so it does not have any effect.