Refactors font loading to be CSP compatible.
The current implementation uses the `onload` attribute to activate lazy-loaded styles. Unfortunately, inlined event handlers like this are not supported by CSP, and the browser will **not** allow them, even if hashed properly.

Instead, I need to pull it out into a separate inline <script /> tag which CSP will allow. Unfortunately, the `load` event does not bubble, so I **cannot** do the more performant:

```javascript
window.addEventListener('load', (evt) => {
  if (evt.target.tagName === 'LINK') evt.target.rel = 'stylesheet';
});
```

This would be more performant because it is a single event listener and a single JavaScript parse/execution. Since `load` doesn't bubble, the event listener never triggers.

Instead, we need to generate a `<script />` tag for every font loaded, which creates the `<link />` via JavaScript and attaches the `onload` event listener. This is far more complicated and less performant than it needs to be, but this is what I need to do in order to be compatible with CSP.