Injects a content security policy into HTML files.
This runs the CSP injector to generate a policy for all HTML files. Since all are generated from source code with no user input, this should be a safe and reasonable thing to do.

CSP injection is executed after HTML minification, because inline JS may also be minified and CSP needs to generate a hash based on the final script source.

We're also manually setting 'self' and an additional hash on the `script-src` to support live reload with 11ty. That script gets injected after the post-processor, so we can't compute the hashes for it normally. These values are included on all builds, which means even production users have unnecessary data here. I'll have to follow up with a dev/prod build distinction to remove them from the prod build while still support live reload for development uses.