Adds environment-specific builds.
This introduces `build:dev` and `build:prod` NPM scripts (as well as `start:*` equivalents). This determines whether or not the build is for development or production and sets an 11ty data variable `environment` to match. This can be used in Nunjucks templates to conditionalize logic.

I implemented most of this into a separate source file from the 11ty data. This allows it to be imported and referenced elsewhere in the build while still being available for Nunjucks templates.

CI will build and deploy the production build to the `main` and `deploy` branches. Post previews will use `dev` builds however to easy debuggability.

I opted to use a custom environment variable `DWAC_ENV` to determine which environment is used. I could use `NODE_ENV`, however it appears to simply be convention, with no strongly defined semantics. Other code in the build may also read that value and change their behavior, which is not necessarily desirable. There is no need for production builds to necessarily be more performant that dev builds. For now, I'll just use `DWAC_ENV`, and we can revist other environment variables if they become useful later.