Disables Rollup optimizations for dev builds.
This reduces build time and improves debuggability.

Unfortunately, there is no easy way to reuse the `environment.ts` file as it is in TypeScript and not easily compiled before executing Rollup. Instead I just manually checked the same environment variable. This is a little redundant, but it is the simplest solution to this problem.