Adds 404 page with handles.
While this page exists at `/404/`, it is served for any unknown path with a 404 status code.

This requires two configurations to work. In Netlify, the redirect config makes all unknown requests serve the 404 page (while it's called a "redirect", in this case it does not actually redirect). In Eleventy, Browsersync has new middleware for serving the 404 page. This makes local builds behave just like the production environment.

I was too lazy to factor out the Browsersync part into its own TypeScript file. It's not likely to grow, so this should be good enough for the time being.

For now, the 404 page just has placeholder content. Structure and style will come later.