Adds `oneline` filter.
This takes a multiline string and removes all the line breaks, replacing them with single (deduplicated) spaces. It effectively takes a paragraph of text and collapses it into a single line.

Main desire for this is because the index page description was long enough to require multiple lines, but this would persist into the final HTML file, since the optimizers can't remove the line breaks without changing semantics. Since this property just exists for SEO, it is safe to remove the line breaks.