Adds feature policy.
This updates the Netlify configuration to include the most restrictive feature policy currently supported.

I went down the list of all features supported by at least one browser and not protected by a flag. None of these features are actually used, so they can all be excluded.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy

I couldn't find a way to restrict this header to just HTML files. Conditionals don't seem to have the power to match all `text/html` documents, and we don't require `*.html` extensions to match on. Instead I just added the header to all resources. This is a bit redundant and wastes bits, but I don't see a better way of doing it.

https://community.netlify.com/t/setting-response-headers-only-on-documents/6144/28

Unfortunately this feature policy is just hard-coded in the Netlify configuration, so this list cannot be easily generated or shared with a similar implementation in the dev build. For now, I'll just leave it here as that should be good enough, but in the future we may want to generate this configuration file to share feature policy data with the 11ty build.