Adds preconnect ot Google Fonts CDN.
The Google Fonts stylesheet returns a link to the font URL which is hosted on the gstatic CDN. By preconnecting to that CDN, we're able to accelerate font loading by resolving DNS and the TLS handshake before the font stylesheet has been downloaded.

I opted to place this after the non-blocking `<link />` tag is generated so as to let that network request come first and take priority. However the preconnect comes before the blocking `<noscript />` link so both requests can be made in parallel in that situation. I'm not sure if this is necessarily the best strategy, but it's my best guess until someone tells me otherwise.