Adds an attribution page to explicitly given copyright and license notice for tools and systems used by the website.
For now, this just includes dependencies visible at runtime (LitElement, lit-html, and TypeScript). All of these already have `@license` comments in the output bundle, so I'm not totally sure if this is strictly necessary, but it should err on the side of safety.

In the future, I'd like to properly attribute all dev and transitive dependencies as well, but that is too much to do by hand. I'll need to look into tooling to generate that list.