Adds copyright notice and link to new attributions page to the footer.
The copyright is generated based on the build date, which should be accurate enough for this blog. It won't update immediately when the year changes, but will fix itself on the next release.

I opted to interpret the year as UTC. Since this is generated at build time I can't make it specific to the user (without introducing JavaScript, which I don't want to do). I could make it use PST because that's where I (as the blog's owner) live. However, even that would vary based on daylight savings, and a deployment build could happen anywhere globally.

Ultimately the year needs to change at one particular instant, so I need to prefer some timezone in order to make that happen. The simplest option is to just use UTC time, as it is the closest I can get to a universal "source of truth".