Adds logo to the header.
Pulled this logo from Wikipedia, compressed it and updated the color: https://en.wikipedia.org/wiki/Raised_fist#/media/File:Fist_.svg

For now, we're just loading it via an `<img />` tag, so unfortunately we can't take advantage of CSS manipulation. For now, it is always white, so that should be good for the time being. I'd rather keep it as a separate file so it can be cached between pages.

Compressed the icon with https://jakearchibald.github.io/svgomg/, because Jake makes all the tools I need.

Also added it to the attribution page, per Creative Commons requirements.