Adds favicon.
This takes the logo and repurposes it as a favicon. I mostly followed the advice from here: https://css-tricks.com/svg-favicons-and-all-the-fun-things-we-can-do-with-them/.

The SVG is a fork of the logo with a blue color. It also includes a dark theme color which apparently works in icons.

I converted the SVG to an ICO file with https://cloudconvert.com/svg-to-ico. I had to manually edit in the transparency with https://www.gimp.org/. This ICO file is used as a fallback for older browsers. I didn't bother with the Safari-specific, pinned icon stuff for now. I'll handle Safari stuff later.

This icon is added on all pages, and makes me want to make a template for all of these pages to extend which includes this structure and makes it easier to maintain.