Adds spacing between logo and header.
In production builds, whitespace is removed which pulls these two elements slightly closer to each other than desired. This additional margin spaces them out as expected.

This does mean that dev and prod builds render slightly differently in this regard, though I don't see a good way of styling this such that it will display the same whether the whitespace is collapsed or not. I'll deal with that later if it becomes significant enough for me to care.