Adds parameters to footer for passing missing them.
The footer macro does not inherit global variables like regular pages do. So these values need to be explicitly passed in. This was done on the index page, but forgotten for others, causing bad links to be generated.

For the case of `post.njk`, it imports a separate `author()` macro, which collides with the global `author` object. Renamed this to `authorFrag` to disambiguate.