Sets `throwOnUndefined` in Nunjucks config.
This makes rendering an `undefined` or `null` value an error in Nunjucks. Enabling this option successfully catches the bug fixed in the prior commit by failing when `author` or `repo` are not provided (or given a bad value, such as a fragment render function, rather than an object with specific properties).

I wasn't able to find a way to explicitly define that a particular macro's parameters are required, however this should have a similar effect.