Readding `trim` and `safe` filters to font.
Turns out the trim and safe were necessary for the JavaScript portion. I think my mistake was that live reload doesn't seem to take macros into account. While it detects the change and refreshes the page, it does not actually use the updated macro. So I believed this was working, when it actually was not.