Adds base template.
This provides the standard frame of the web page, including the header, footer, and linking fonts, styles, metadata, etc.

I'm working to migrate all pages to use this template. The last few bugs I've fixed have largely come from the fact that work done on one page was not reflected on all other pages. Hopefully by using this template, those pages should be consistent.

Most of the actual content here was forked from the index page. I updated classes to be namespaced to ensure that they don't conflict with those from other macros on the page.

I have no doubt merging all pages to use this template will actually introduce layout and styling bugs in the short term, but once those are addressed, in the long term they should be much more consistent with each other.

Unfortunately there seems to be a weird bug with using a `block` inside a `set`. It just doesn't work if the block isn't provided which means that we can't conditionalize based on whether or not a block is provided. This is particularly annoying for including JavaScript, since not all pages require it. Instead I had to add a generic `head` block for adding any arbitrary elements into the `<head />` tag, when really I wanted a simple `script` block that works just like CSS except for JavaScript. I don't see a good way of implementing that pattern right now, so this work around is a good as I can do.