Migrates `/deployment` to use new base template.
This isn't totally necessary since it is just a debug page, but this is good to bring it into alignment with the rest of the pages.

I also frequently look at this page in the mobile emulator, since that just happens to be what I'm inspecting with, and by using this template it implicitly uses the viewport `<meta />` tag which makes this page much more legible in the emulator. It still looks pretty ugly, but there's no need for it to be pretty right now.