Enlarges share icons.
Previously, share icons were only 24px tall, which can be hard to accurately touch on a mobile device. Recommendations vary, but a simple rule of thumb is that it should be at least 44px x 44px. https://medium.com/@zacdicko/size-matters-accessibility-and-touch-targets-56e942adc0cc#:~:text=In%20the%20Human%20Interface%20Guidelines,be%20trying%20to%20tap%20successfully.

I scaled up the images to match that and spaced them out/line wrapped them to look aesthetically pleasing. These images are not square and only the height is fixed to 44px, so technically the clipboard icon is below the minimum width. I'll probably need to expand the tap target beyond the image dimensions to look good and still meet the touch target requirements, but I'll come to that later.