Adds timestamp to posts.
This allows `::: timestamp :::` syntax in Markdown files which will generate HTML containing the `date` value in the front matter of the post. This is appropriately styled and looks best immediately under the first header of the post.

This required manually setting the Markdown library rather than using 11ty's defaults. Ultimately this is the same library with the same configuration, so this shouldn't really change anything aside from the actual `::: timestamp :::` syntax.

Unfortunately, 11ty appears to infer the `date` value from the "Created By" property of the file on the file system. This means there's no good way to validate that a `date` value was explicitly provided by the author as I would want to do. I still validate it just for good typing, but I don't expect it to ever actually throw for that condition.