Adds TypeScript syntax highlighting theme.
Mostly based this off Visual Studio code to get something that looks reasonably ok.

For now, languages are loaded via `all.css` which will include all language themes. Not every page will need this, so at some point in the future it would be nice to have a page declare which themes it needs and generate `<link />` tags appropriately to load the minimum amount.