Simplifies TypeScript syntax highlighting theme.
This changes `[class="language-typescript"]` to `.language-typescript`, which is more shorter and more intuitive. I was originally using `[class*="language-"]` before deciding to make it specific to TypeScript. I just didn't realize the syntax could be shortened further.

Also removed the `.token` class. Since everything is namespaced to `code.language-typescript` anyways, the `token` class is not needed.