Adds Plausible script.
I set up an account with the current prod URL (dwac.netlify.app) and copy-pasted the script into the page template which should be applied to every page.

The script itself is pretty minimal and is property asynchronous. It is supposed to include a `data-domain="dwac.netlify.app"` attribute, however it seems to infer the host origin if not specified. I don't want builds to be conditionalized on the environment they are being deployed to, so I would rather not put the origin in the actual source code. Letting Plausible infer from the current origin seems like the right behavior for 11ty on Netlify.

The one significant limitation I see is that custom events do not allow any special metadata. In particular, this means I can't track scroll depth on a page, which is quite dissapointing. I could potentially hack around this with explicit events for "quarter scroll", "half scroll", "full scroll", etc. For now, all this tracks is page views, which is a reasonable starting point. See: https://github.com/plausible/analytics/issues/134.