Removes `blog-next.dwac.dev`.
Apparently Netlify's domain aliases feature only aliases the **production** domain, so this would always resolve to the production deployment and not the one from the `main` branch.

I don't see a good way of supporting this, so I'm just removing the `blog-next.dwac.dev` domain. Unfortunately, this means that Google Domains cannot be verified on a non-prod domain, but hopefully that should not be a common need.