Shortens page title onto oneline.
This trims whitespace and removes newlines that won't show in page titles anyways.

Also updated the description to filter when the variable is set, so it can be reused without having to filter again.

The same can't be done for `styles` and `content`, becauseh the formers uses an async filter (`css`) and both use `safe` which doesn't seem to work in a block context.