Adds email to home page.
I was able to set up an email at this domain and verified that it works. This change is minimal, but does expose the email in a very easily spammable way. There are techniques to make email addresses less scrapable, but they have significant usability or accessibility issues. Ultimately there's no way to hide this such that a `mailto:` link would work but scrapers won't find it without using client-side JS, which wouldn't work for `<noscript />` users and would be much less accessible.

Other techniques that use a proxy email address would still forward to my inbox anyways. Instead I'm just not going to try and accept that spam email will go to this address and rely on spam filters instead.