Switches `style-src` CSP to use 'unsafe-inline'.
Inline styles of Lit Element were being blocked in Firefox because it does not support adopting style sheets. This means Lit Element falls back to appending a `<style />` tag which requires `unsafe-inline` to work.

Unfortunately, on modern browsers, `unsafe-inline` is ignored when explicit hosts/hashes are given for backwards compatibility reasons. This means we can't use `unsafe-inline` *and* provide explicit hosts. Instead, the best we can do is `style-src 'unsafe-inline' https:` which will work in Firefox although it will lessen security on Chrome.