Updates dependencies via `npm update`.
This addresses a security vulnerability with `socket.io` and also updates everything else just because.

This also updated `browser-sync` which broke the CSP hash I had hard-coded. This is unfortunately a known issue with no good workaround. Updated CSP logic to compute the hash directly from source content so it will be easier to update in the future.