Update `src/client/components/share.ts` to explicitly check for the existence of `navigator` APIs.
This is necessary to avoid a breakage in TS 4.2 from [improved uncalled function checks](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-4-2.html#improved-uncalled-function-checks-in-logical-expressions).

`navigator.share` and `navigator.clipboard` are always present according to types, but may not actually exist on real browsers, since some don't. For example, desktop browsers don't have implementations for `navigator.share` and leave it `undefined`.

The fix here is to add a `!!` to the expression to more clearly test that it exists and make TypeScript treat the expression as a boolean rather than an uncalled function.