Inlines Lora font in the base template.
Current font best practices are to inline fonts and self-host them to reduce TLS handshakes and overal network requests. The Lora CSS file is now inlined into the compiled and minified styles for each page that uses the base template (all of them). The CSS file references locally hosted `woff2` font files and preloads the `latin.woff2` as that is almost certainly required. `latin-ext` is hosted, but not preloaded as it most likely will not be used at all, and even if so, it probably won't be visible on the initial viewport.

I opted to use `font-display: fallback;` as that seems to be the current best practice, so the font is downloaded and blocks for a short time before giving up and using the fallback fonts. Repeated page load should be cached however and use the font immediately. I'll need to do some prod testing to verify this, but blocking the font in local testing seems to work as expected. `font-display: optional;` is used as the fallback behavior for older browsers as I'm not able to confirm browser support for `fallback`, but I believe it is a relatively new value.